# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TRIXHUB Affiliation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me/dashboard:
    get:
      operationId: getDashboard
      tags: [users]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark task as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Already completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/me/preferred-currency:
    patch:
      operationId: updatePreferredCurrency
      tags: [users]
      summary: Update preferred currency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCurrencyBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    UpdateProfileBody:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
    UpdateCurrencyBody:
      type: object
      properties:
        currency:
          type: string
      required: [currency]
    User:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        isActivated:
          type: boolean
        referralCode:
          type: string
        referredByCode:
          type: ["string", "null"]
        preferredCurrency:
          type: string
        createdAt:
          type: string
      required: [id, firstName, lastName, email, phone, country, isActivated, referralCode, preferredCurrency, createdAt]
...
    Dashboard:
      type: object
      properties:
        totalBalance:
          type: number
        referralBalance:
          type: number
        taskBalance:
          type: number
        withdrawnAmount:
          type: number
        spentAmount:
          type: number
        inactiveBalance:
          type: number
        totalReferrals:
          type: integer
        activeReferrals:
          type: integer
        inactiveReferrals:
          type: integer
        level1Count:
          type: integer
        level2Count:
          type: integer
        level3Count:
          type: integer
        pendingWithdrawals:
          type: integer
        currency:
          type: string
        exchangeRate:
          type: number
      required: [totalBalance, referralBalance, taskBalance, withdrawnAmount, spentAmount, inactiveBalance, totalReferrals, activeReferrals, inactiveReferrals, level1Count, level2Count, level3Count, pendingWithdrawals, currency, exchangeRate]
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        reward:
          type: number
        type:
          type: string
        isCompleted:
          type: boolean
        completedAt:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
      required: [id, title, description, reward, type, isCompleted]
...